Add completion message and check for existence of directories
git-svn-id: http://svn.opengeo.org/suite/trunk@1120 fe66add1-021f-4af3-b75b-ea5154e73f91

+++ installer/mac/binaries/scripts/suite_uninstall.sh
@@ -22,6 +22,12 @@ if [ -f /private/etc/manpaths.d/opengeo-pgsql ]; then
 fi
 
 # Remove receipts entries
-find /Library/Receipts -name "org.opengeo.*" -exec /bin/rm -rf {} ';'
-find /var/db/receipts -name "org.opengeo.*" -exec /bin/rm -rf {} ';'
+if [ -d /Library/Receipts ]; then
+  find /Library/Receipts -name "org.opengeo.*" -exec /bin/rm -rf {} ';'
+fi
+if [ -d /var/db/receipts ]; then
+  find /var/db/receipts -name "org.opengeo.*" -exec /bin/rm -rf {} ';'
+fi
+
+echo "The OpenGeo Suite is now uninstalled."
